third_party/WebKit/LayoutTests/svg/dom/SVGMarkerElement-orientType-synchronization.html

<!DOCTYPE html>
<title>SVGMarkerElement.orientType synchronization</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var marker = document.createElementNS("http://www.w3.org/2000/svg", "marker");
    assert_equals(marker.orientType.baseVal, SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE, "initial");

    marker.orientType.baseVal = SVGMarkerElement.SVG_MARKER_ORIENT_AUTO;
    assert_equals(marker.getAttribute("orient"), "auto");
    assert_equals(marker.orientType.baseVal, SVGMarkerElement.SVG_MARKER_ORIENT_AUTO, "to 'auto'");
});
</script>
